+++ dagger/dagger.rs
@@ -1,7 +1,5 @@
 // Copyright 2022-2024 Protocol Labs
 // SPDX-License-Identifier: Apache-2.0, MIT
-use std::sync::Arc;
-
 use color_eyre::eyre::{bail, Result};
 use dagger_sdk::{
     logging::StdLogger, Container, ContainerWithDirectoryOptsBuilder,
@@ -9,6 +7,11 @@ use dagger_sdk::{
     ContainerWithFileOptsBuilder, DaggerConn, Directory, HostDirectoryOpts,
 };
 use fs_err as fs;
+use std::future::Future;
+use std::sync::Arc;
+
+mod minimal;
+use self::minimal::*;
 
 /// Execute the lazily prepared container definition and convert to a `Result`
 async fn run(container: &Container) -> Result<()> {
@@ -110,133 +113,67 @@ impl WithCache for Container {
     }
 }
 
-/// Create a container definition which is able to compile the contracts
-fn define_contracts_container(client: DaggerConn) -> Result<Container> {
-    let compiled_contracts_dir = "/workdir/compiled_contracts";
-
-    let container = client.container()
-        .from("docker.io/library/node:latest")
-    .with_mounted_directory("/workdir", hrrd(&client))
-    .with_mounted_directory(compiled_contracts_dir, hccd(&client))
-    .cache_apt(&client)
-    .with_exec(cmd("apt-get update -y"))
-    .with_exec(cmd("apt-get install -y curl which"))
-    .with_workdir("/workdir/contracts")
-    .with_exec(cmd("ls -al"))
-    .cache_node(&client)
-    .with_exec(cmd("npm install -g pnpm"))
-    .with_exec(cmd("pnpm install"))
-    .with_exec(vec!["sh", "-c", "curl -L https://foundry.paradigm.xyz | bash && /root/.foundry/bin/foundryup --install 0.3.0"])
-    .with_exec(cmd("git submodule update --init --recursive"))
-    .with_exec(cmd(format!("mkdir -p {compiled_contracts_dir}")))
-    .with_env_variable_opts("PATH", "${PATH}:/root/.foundry/bin", ContainerWithEnvVariableOptsBuilder::default().expand(true).build()?)
-    .with_exec(cmd("which forge"))
-    // actually build the contracts
-    // TODO investigate caching issue further
-    .with_exec(cmd(format!("forge -vvv build -C ./src/ --lib-paths ./lib/ --via-ir --sizes --skip test --out={compiled_contracts_dir}")));
-
-    Ok(container)
-}
-
-fn define_crates_container(client: DaggerConn) -> Result<Container> {
-    let container = client
-        .container()
-        .from("docker.io/rust:bookworm")
-        .with_mounted_directory("/workdir", hrrd(&client))
-        .with_mounted_directory("/workdir/_compiled_contracts", hccd(&client))
-        .with_workdir("/workdir")
-        .cache_apt(&client)
-        .with_exec(cmd("apt-get update"))
-        .with_exec(cmd(
-            "apt-get install -y build-essential clang cmake protobuf-compiler",
-        ))
-        .cache_rust(&client)
-        // actually build the rust binaries
-        .with_exec(cmd("cargo b -p fendermint_app -p ipc-cli"))
-        // see what was created
-        .with_exec(cmd("ls -al /workdir/target/debug"))
-        .with_exec(cmd("ls -al output"));
-
-    Ok(container)
-}
-
 fn hccd(client: &DaggerConn) -> Directory {
     dir(client, "fendermint/actors/output")
 }
 
-async fn prepare_fendermint_two_stage_build(client: DaggerConn) -> Result<Container> {
-    fs::create_dir_all("_compiled_contracts")?;
+#[derive(clap::Parser, Debug, Clone)]
+#[command(version, about, long_about = None)]
+struct Arg {
+    #[command(subcommand)]
+    what: Option<What>,
+}
 
-    let contracts_out = dir(&client, "./contracts/out");
-    let node_app_config = dir(&client, "./fendermint/app/config");
+#[derive(Debug, Clone, Default, clap::Subcommand)]
+enum What {
+    #[default]
+    #[command()]
+    All,
 
-    let contracts_gen = define_contracts_container(client.clone())?;
-    run(&contracts_gen).await?;
+    #[command()]
+    MinimalContainer,
 
-    let crates_def = define_crates_container(client.clone())?;
-    run(&crates_def).await?;
+    #[command()]
+    CargoTests,
+}
 
-    let f_fendermint = crates_def.file("/workdir/target/debug/fendermint");
-    let f_ipc = crates_def.file("/workdir/target/debug/ipc-cli");
+use std::pin::Pin;
 
-    let car_extra =
-        contracts_gen.file("/workdir/fendermint/actors/output/custom_actors_bundle.car");
-
-    // prepare the to-be-published "runner" container
-    let runner = client
-        .container()
-        .from("docker.io/debian:bookworm-slim")
-        .with_file_opts(
-                "/usr/local/bin/fendermint",
-            f_fendermint,
-            ContainerWithFileOptsBuilder::default()
-                .permissions(0o755_isize)
-                .build()?,
-        )
-        .with_file_opts(
-            "/usr/local/bin/ipc-cli",
-            f_ipc,
-            ContainerWithFileOptsBuilder::default()
-                .permissions(0o755_isize)
-                .build()?,
-        )
-        .with_exec(cmd("ls -al"))
-        .cache_apt(&client)
-        .with_exec(vec!["sh", "-c", "apt-get update && \
-            apt-get install -y libssl3 ca-certificates curl"])
-        .with_env_variable("FM_HOME_DIR", "/fendermint")
-        .with_exposed_port(26658)
-        .with_exposed_port(8445)
-        .with_exposed_port(9184)
-        .with_entrypoint(cmd("docker-entry.sh"))
-        .with_default_terminal_cmd(cmd("run"))
-        // TODO STOPSIGNAL SIGTERM
-        .with_env_variable("FM_ABCI__LISTEN__HOST", "0.0.0.0")
-        .with_env_variable("FM_ETH__LISTEN__HOST", "0.0.0.0")
-        .with_env_variable("FM_METRICS__LISTEN__HOST", "0.0.0.0")        
-        .with_exec(cmd("mkdir -p /fendermint/logs"))
-        .with_exec(cmd("chmod 777 /fendermint/logs"))
-        .with_file("/usr/local/bin/docker-entry.sh", hrrd(&client).file("fendermint/docker/docker-entry.sh"))
-        .with_file("/fendermint/custom_actors_bundle.car", car_extra)
-        // TODO FIXME - this is insane
-        .with_exec(cmd("mkdir -p /fendermint/builtin-actors/output"))
-        .with_exec_opts(cmd("curl -L -o /fendermint/builtin-actors/output/bundle.car https://github.com/filecoin-project/builtin-actors/releases/download/${BUILTIN_ACTORS_TAG}/builtin-actors-mainnet.car"), ContainerWithExecOptsBuilder::default().expand(true).build()?)
-        .with_directory("/fendermint/contracts", contracts_out)
-        .with_directory_opts("/fendermint/config", node_app_config, ContainerWithDirectoryOptsBuilder::default().exclude(vec![".git", ".gitignore", ".*"]).build()?);
-
-    run(&runner).await?;
-
-    // TODO extract file and publish gh release IFF tagged
-    // TODO
-    // TODO runner.publish(address).await?;
-
-    Ok(runner)
+impl What {
+    pub fn into_fut(
+        self,
+    ) -> Box<dyn FnMut(DaggerConn) -> Pin<Box<dyn Future<Output = Result<()>>>>> {
+        match self {
+            Self::All => Box::new(move |client: DaggerConn| {
+                Box::pin(async move {
+                    // TODO launch in parallel
+                    prepare_fendermint_two_stage_build(client.clone()).await?;
+                    run_cargo_test(client).await?;
+                    Ok(())
+                })
+            }),
+            Self::MinimalContainer => Box::new(move |client: DaggerConn| {
+                Box::pin(async move {
+                    prepare_fendermint_two_stage_build(client).await?;
+                    Ok(())
+                })
+            }),
+            Self::CargoTests => Box::new(move |client: DaggerConn| {
+                Box::pin(async move {
+                    run_cargo_test(client).await?;
+                    Ok(())
+                })
+            }),
+        }
+    }
 }
 
 #[tokio::main]
 async fn main() -> Result<()> {
     color_eyre::install()?;
     // dagger_sdk::logging::default_logging()?;
+
+    let args = <Arg as clap::Parser>::try_parse()?;
 
     dagger_sdk::connect_opts(
         dagger_sdk::Config {
@@ -246,13 +183,9 @@ async fn main() -> Result<()> {
             execute_timeout_ms: None,
             logger: Some(Arc::new(StdLogger::default())),
         },
-        |client| async move {
-            let _fendermint_featherweight =
-                prepare_fendermint_two_stage_build(client).await.unwrap();
-
-            // let anvil = prepare_anvil_service(client).await?;
-            // let client = prepare_ipc_cli(client).await?;
-
+        move |client| async move {
+            let fx = args.what.unwrap_or_default().into_fut()(client);
+            fx.await?;
             Ok(())
         },
     )
